move convert to new abstract class
